+++ venv.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-# Usage: source venv.sh
-source .venv/bin/activate